Reoroganize VS Dash utils
