+++ frontend/src/utils/dateUtils.ts
@@ -30,3 +30,40 @@ export function toEST(date: string) {
   });
   return etDate;
 }
+
+export function formatShortDate(
+  dateInput: string | Date | null | undefined
+): string {
+  if (!dateInput) return '';
+  const dateObj = new Date(dateInput);
+  if (Number.isNaN(dateObj.getTime())) return '';
+  return dateObj.toLocaleDateString('en-US', {
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric'
+  });
+}
+
+export function formatRange(
+  start?: string | Date | null | undefined,
+  end?: string | Date | null | undefined
+): string {
+  const startStr = formatShortDate(start);
+  const endStr = formatShortDate(end);
+  if (!startStr && !endStr) return 'No Dates Available';
+  // Ensures only end date is shown per CRASM-3140
+  if (startStr && endStr) return `${endStr}`;
+  return startStr || endStr;
+}
+
+export function enrolledWithinTwoWeeks(timestamp?: string | null): boolean {
+  if (!timestamp) return false;
+
+  const enrolledDate = new Date(timestamp);
+  if (isNaN(enrolledDate.getTime())) return false;
+
+  const now = new Date();
+  const twoWeeksInMs = 14 * 24 * 60 * 60 * 1000;
+
+  return now.getTime() - enrolledDate.getTime() <= twoWeeksInMs;
+}
